Refactor: api 로직을 Follower에서 받아오도록 수정 (#66)
