+++ src/pages/Profile/Follower/index.jsx
@@ -1,12 +1,13 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
+import axios from 'axios';
+
 import BaseHeader from '../../../components/common/BaseHeader';
 import FollowerItem from '../../../components/FollowerItem';
 
 import arrowIcon from '../../../assets/icon/icon-arrow-left.png';
 import { IR } from '../../../styles/Util';
-import dummyList from '../../../components/FollowerItem/dummyList';
 
 const SContainer = styled.section`
   padding: 24px 16px 0;
@@ -23,8 +24,37 @@ function Follower() {
     navigate('/profile');
   };
 
+  const [followerData, setFollowerData] = useState([]);
+
+  const fetchFollowerList = async () => {
+    try {
+      const response = await axios.get(
+        `https://mandarin.api.weniv.co.kr/profile/${JSON.parse(
+          localStorage.getItem('accountName')
+        )}/follower`,
+        {
+          headers: {
+            Authorization: `Bearer ${JSON.parse(
+              localStorage.getItem('token')
+            )}`,
+            'Content-type': 'application/json',
+          },
+        }
+      );
+      setFollowerData(response.data);
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
+  useEffect(() => {
+    fetchFollowerList();
+  }, []);
+
+  // console.log(followerData, followerData.length);
+
   return (
-    <div>
+    <>
       <BaseHeader
         leftIcon={arrowIcon}
         leftClick={handleToProfile}
@@ -32,9 +62,12 @@ function Follower() {
       />
       <SContainer>
         <STitle>팔로워 페이지</STitle>
-        <FollowerItem />
+        {followerData.length > 0 &&
+          followerData.map((data) => (
+            <FollowerItem key={data.accountname} data={data} />
+          ))}
       </SContainer>
-    </div>
+    </>
   );
 }
